+++ src/test/java/learn/agile/demos/pages/DemoPage.java
@@ -46,9 +46,9 @@ public class DemoPage {
 	 * @param x
 	 * @param y
 	 */
-	public void multiplyNumbers(final int x, final int y) {
-		driver.findElement(By.id("first-number")).sendKeys(Integer.toString(x));
-		driver.findElement(By.id("second-number")).sendKeys(Integer.toString(y));
+	public void multiplyNumbers(final float x, final float y) {
+		driver.findElement(By.id("first-number")).sendKeys(Float.toString(x));
+		driver.findElement(By.id("second-number")).sendKeys(Float.toString(y));
 		driver.findElement(By.id("multiplyer-button")).click();
 	}
 
@@ -56,10 +56,10 @@ public class DemoPage {
 	 * @return add results
 	 * @throws InterruptedException
 	 */
-	public int getCalculatorResults() throws InterruptedException {
+	public float getCalculatorResults() throws InterruptedException {
 		WebElement resultEl = (new WebDriverWait(driver, 300))
 				.until(ExpectedConditions.presenceOfElementLocated(By.id("calculator-result")));
 		Thread.sleep(100); // timing - for some reason on some OS's webdriver returns too quickly;
-		return Integer.parseInt(resultEl.getText());
+		return Float.parseFloat(resultEl.getText());
 	}
 }
